# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Clawdbot/Ally API
  description: |
    AI-powered personal assistant API for building conversational experiences,
    managing user context with persistent memory, and extending functionality
    through a modular skills system.
    
    ## Authentication
    
    All API requests require authentication via one of:
    - **API Key:** `Authorization: Bearer <api_key>`
    - **JWT Token:** `Authorization: Bearer <jwt_token>`
    - **OAuth2:** Standard OAuth2 bearer token
    
    ## Rate Limits
    
    Rate limits vary by plan tier. See the Rate Limits section for details.
  version: 1.0.0
  contact:
    name: Clawdbot API Support
    email: api-support@clawdbot.com
    url: https://docs.clawdbot.com
  license:
    name: Proprietary
    url: https://clawdbot.com/terms
  termsOfService: https://clawdbot.com/terms
servers:
  - url: https://api.clawdbot.com/v1
    description: Production server
  - url: https://api-staging.clawdbot.com/v1
    description: Staging server
  - url: http://localhost:3001/api
    description: Local development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /provision:
    post:
      summary: Provision new agent
      description: |
        Manually trigger agent provisioning.
        
        Requires admin API key authentication.
      tags: [Agents]
      operationId: provisionAgent
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, tier, email]
              properties:
                customer_id:
                  type: string
                  description: Customer UUID
                tier:
                  type: string
                  enum: [starter, professional, enterprise]
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Provisioning started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deployment_id:
                    type: string
                  status:
                    type: string
                  request_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /provision/status/{deployment_id}:
    get:
      summary: Get provision status
      description: Check the status of an ongoing provisioning
      tags: [Agents]
      operationId: getProvisionStatus
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, in_progress, completed, failed]
                  progress:
                    type: integer
                    minimum: 0
                    maximum: 100
                  logs:
                    type: string
                  error:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /provision/list:
    get:
      summary: List deployments
      description: List all deployments for the account
      tags: [Agents]
      operationId: listDeployments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Deployments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        agent_id:
                          type: string
                        status:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  count:
                    type: integer

  # ===================
  # WEBHOOKS
  # ===================
  /webhooks/stripe:
    post:
      summary: Stripe webhook
      description: |
        Receive Stripe webhook events.
        
        Events handled:
        - `checkout.session.completed` → Provision new agent
        - `customer.subscription.updated` → Update tier
        - `customer.subscription.deleted` → Suspend agent
        - `invoice.payment_failed` → Handle payment failure
      tags: [Webhooks]
      operationId: handleStripeWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook event payload
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key for server-to-server communication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token or API key
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          type: object
          properties:
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request body
              details:
                field: message
                issue: Required field missing
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded. Please try again in 60 seconds.
              details:
                limit: 100
                remaining: 0
                reset_at: "2026-01-27T10:31:00Z"
...